Basic ESPINA gui
Displays the views in the sammer manner than ESPINE does.
At the moment, only some basic 2D views are used without
configuration.

SliceWidget represents the widget containing both the view
and the controls to iterate over it

VolumeWidget nos is just a copy of SliceWidget intended to be
replaced in a future by the widget with the 3D view and its
controls.